YOLFin/requirements.txt
pandas
ccxt 
datetime
os
matplotlib
backtrader
pyts
pillow